+++ nutshell/compiler.py
@@ -35,11 +35,12 @@ def _iter_transitions(tbl):
             seen.add(tr.ctx)
             lno, start, end = tr.ctx
             start, end = None if not start else start - 1, None if end is None else end - 1
-            if src:
-                yield from ('', src.format(line=lno+tbl.start, span=tbl[lno-1][start:end]))
             if cmt:
-                yield from (tbl.comments.pop(cmt_lno) for cmt_lno in list(tbl.comments) if cmt_lno < lno)
-            if lno in tbl.comments:
+                yield from [tbl.comments.pop(cmt_lno) for cmt_lno in list(tbl.comments) if cmt_lno < lno]
+            if src:
+                # yield ''
+                yield src.format(line=lno+tbl.start, span=tbl[lno-1][start:end])
+            if cmt and lno in tbl.comments:
                 yield '{}{}'.format(', '.join(map(str, tr)), tbl.comments.pop(lno))
                 continue
         yield ', '.join(map(str, tr))